+++ packages/replay_bloc/test/blocs/counter_bloc.dart
@@ -1,38 +1,56 @@
 import 'package:bloc/bloc.dart';
 import 'package:replay_bloc/replay_bloc.dart';
 
-enum CounterEvent { increment, decrement }
+abstract class CounterEvent extends ReplayEvent {}
+
+class Increment extends CounterEvent {}
+
+class Decrement extends CounterEvent {}
 
 class CounterBloc extends ReplayBloc<CounterEvent, int> {
-  CounterBloc({int limit}) : super(0, limit: limit);
+  CounterBloc({
+    int limit,
+    this.onEventCallback,
+    this.onTransitionCallback,
+  }) : super(0, limit: limit);
+
+  final void Function(ReplayEvent) onEventCallback;
+  final void Function(Transition<ReplayEvent, int>) onTransitionCallback;
 
   @override
   Stream<int> mapEventToState(CounterEvent event) async* {
-    switch (event) {
-      case CounterEvent.increment:
-        yield state + 1;
-        break;
-      case CounterEvent.decrement:
-        yield state - 1;
-        break;
+    if (event is Increment) {
+      yield state + 1;
+    } else if (event is Decrement) {
+      yield state - 1;
     }
+  }
+
+  @override
+  void onEvent(ReplayEvent event) {
+    onEventCallback?.call(event);
+    super.onEvent(event);
+  }
+
+  @override
+  void onTransition(Transition<ReplayEvent, int> transition) {
+    onTransitionCallback?.call(transition);
+    super.onTransition(transition);
   }
 }
 
-class CounterBlocMixin extends Bloc<CounterEvent, int> with ReplayMixin<int> {
+class CounterBlocMixin extends Bloc<CounterEvent, int>
+    with ReplayBlocMixin<CounterEvent, int> {
   CounterBlocMixin({int limit}) : super(0) {
     this.limit = limit;
   }
 
   @override
   Stream<int> mapEventToState(CounterEvent event) async* {
-    switch (event) {
-      case CounterEvent.increment:
-        yield state + 1;
-        break;
-      case CounterEvent.decrement:
-        yield state - 1;
-        break;
+    if (event is Increment) {
+      yield state + 1;
+    } else if (event is Decrement) {
+      yield state - 1;
     }
   }
 }