+++ packages/replay_bloc/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.1-dev.2
+
+- **BREAKING**: `ReplayMixin` renamed to `ReplayCubitMixin`
+- feat: `ReplayBloc` and `ReplayBlocMixin` support
+
 # 0.0.1-dev.1
 
-- feat: dev release (WIP)
+- feat: `ReplayCubit` and `ReplayMixin` dev release